% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_guane.R
\name{app_guane}
\alias{app_guane}
\title{Assemble the Guane application.}
\usage{
app_guane(module = "full", default_lang = "en")
}
\arguments{
\item{module}{Primary module to launch, or full for all modules.}

\item{default_lang}{Initial interface language: en, es, or pt.}
}
\description{
Assemble the Guane application.
}
