% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_i18n.R
\name{app_i18n}
\alias{app_i18n}
\title{#####################################}
\usage{
app_i18n(path = NULL)
}
\arguments{
\item{path}{path to the translation JSON.}
}
\description{
load dictionaries from json
}
\details{
####################################
}
\examples{
\dontrun{
tr <- app_i18n()
tr$t('signal', 'es')
}

}
\author{
Viviana Romero Alarcon

Read a JSON dictionary self-contained in the package (www folder)
This reproduce the reactive language switching design of   \code{shiny.i18n}
using only \code{jsonlite}.
An object of class \code{app_i18n} with a \verb{$t(key, lang)} translator,
the available \verb{$languages} and the raw \verb{$dict}.
}
