% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Connect the primary module servers.}
\usage{
app_server(modules = c("signal", "asr", "div", "sse"))
}
\arguments{
\item{modules}{Primary modules to connect: signal, asr, div, or sse.}
}
\description{
Connect the primary module servers.
}
