% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Display the Guane interface.}
\usage{
app_ui(
  modules = c("signal", "asr", "div", "sse"),
  default_lang = "en",
  resource_root = guane_resource_root()
)
}
\description{
Display the Guane interface.
}
