% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_read_data.R
\name{core_read_traits}
\alias{core_read_traits}
\title{Read a trait table}
\usage{
core_read_traits(path)
}
\arguments{
\item{path}{Character. Path to a trait table.}
}
\value{
A data frame.
}
\description{
Reads a trait table from common tabular file formats supported by Guane.
}
