% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_read_data.R
\name{core_read_tree}
\alias{core_read_tree}
\title{Read a phylogenetic tree file}
\usage{
core_read_tree(path)
}
\arguments{
\item{path}{Character. Path to a tree file.}
}
\value{
A \code{phylo} or \code{multiPhylo} object.
}
\description{
Reads a phylogenetic tree from common tree formats supported by Guane.
}
