% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_template.R
\name{core_template_summary}
\alias{core_template_summary}
\title{Template core function}
\usage{
core_template_summary(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A list with summary statistics.
}
\description{
This is an example of a core analytical function that can run without Shiny.
}
\keyword{internal}
