% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guane-package.R
\docType{package}
\name{guane-package}
\alias{guane}
\alias{guane-package}
\title{Guane: Multilingual Phylogenetic Comparative Analyses}
\description{
A modular R Shiny platform for data preparation, phylogenetic analyses,
diagnostics, visualizations and reproducible exports in three languages.
}
\author{
\strong{Maintainer}: Leidy Viviana Romero-Alarcon \email{vivianaromeroalarcon@umb.edu}

Authors:
\itemize{
  \item Leidy Viviana Romero-Alarcon \email{vivianaromeroalarcon@umb.edu}
}

}
\keyword{internal}
