% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_mod_signal.R
\name{mod_signal_server}
\alias{mod_signal_server}
\title{Phylogenetic signal module server}
\usage{
mod_signal_server(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny module server.
}
\description{
Phylogenetic signal module server
}
