% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mod_signal.R
\name{mod_signal_ui}
\alias{mod_signal_ui}
\title{Phylogenetic signal module UI}
\usage{
mod_signal_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI.
}
\description{
Phylogenetic signal module UI
}
