% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_mod_template.R
\name{mod_template_server}
\alias{mod_template_server}
\title{Template module server}
\usage{
mod_template_server(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny module server.
}
\description{
Template module server
}
\keyword{internal}
