% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mod_template.R
\name{mod_template_ui}
\alias{mod_template_ui}
\title{Template module UI}
\usage{
mod_template_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI.
}
\description{
Template module UI
}
\keyword{internal}
