% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Guane platform.}
\usage{
run_app(module = "full", default_lang = "en", ...)
}
\arguments{
\item{module}{Primary module to launch, or full for all modules.}

\item{default_lang}{Initial interface language: en, es, or pt.}

\item{...}{Additional arguments passed to shiny::runApp().}
}
\description{
Launch the Guane platform.
}
