% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mod_data_upload.R
\name{ui_mod_data_upload}
\alias{ui_mod_data_upload}
\title{Data upload, preview, and validation module UI}
\usage{
ui_mod_data_upload(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI.
}
\description{
Provides file upload widgets for phylogenetic trees, posterior tree
distributions, and trait datasets, with immediate previews and validation.
}
