% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guane-package.R
\docType{package}
\name{guane-package}
\alias{guane}
\alias{guane-package}
\title{guane: Multilingual Shiny Platform for Phylogenetic Comparative Analyses}
\description{
Guane is a modular Shiny-based platform for phylogenetic comparative analyses. It provides data validation, analytical workflows, multilingual guidance, and reproducible code export for researchers using phylogenetic comparative methods.
}
\author{
\strong{Maintainer}: Viviana Romero-Alarcon \email{vivianaromeroalarcon@umb.edu}

}
\keyword{internal}
