% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Guane Shiny application}
\usage{
run_app(module = "full", ...)
}
\arguments{
\item{module}{Character. Which Guane module to run. Use \code{"full"} for the
complete app or a module name such as \code{"signal"}.}

\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Opens the Guane Shiny application.
}
\description{
Run the Guane Shiny application
}
