% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Guane Shiny application}
\usage{
run_app(module = "data_upload", ...)
}
\arguments{
\item{module}{Module to open. Use "data_upload", "signal", or "template".}

\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\value{
Runs the Shiny app.
}
\description{
Run the Guane Shiny application
}
